[#26] peer, config, and gomo moves under simra directory
