+++ simra/sprite.go
@@ -1,0 +1,7 @@
+package simra
+
+import "github.com/pankona/gomo-simra/simra/peer"
+
+type Sprite struct {
+	peer.PeerSprite
+}
